+++ pkg/plugins/policies/meshaccesslog/plugin/xds/rules.go
@@ -1,8 +1,13 @@
 package xds
 
 import (
+	"fmt"
+	"strconv"
+	"strings"
+
 	envoy_accesslog "github.com/envoyproxy/go-control-plane/envoy/config/accesslog/v3"
 
+	common_api "github.com/kumahq/kuma/v2/api/common/v1alpha1"
 	bldrs_accesslog "github.com/kumahq/kuma/v2/pkg/envoy/builders/accesslog"
 	. "github.com/kumahq/kuma/v2/pkg/envoy/builders/common"
 	"github.com/kumahq/kuma/v2/pkg/plugins/policies/core/rules/inbound"
@@ -10,6 +15,28 @@ import (
 	"github.com/kumahq/kuma/v2/pkg/util/pointer"
 	listeners_v3 "github.com/kumahq/kuma/v2/pkg/xds/envoy/listeners/v3"
 )
+
+// MatchToCEL converts a Kuma inbound Match into a CEL expression evaluating to
+// true when the Envoy connection satisfies all sub-matchers in the Match.
+// Returns "" for a nil/empty Match (which conceptually matches all traffic).
+func MatchToCEL(m *common_api.Match) string {
+	if m == nil {
+		return ""
+	}
+	var parts []string
+	if m.SpiffeID != nil {
+		switch m.SpiffeID.Type {
+		case common_api.ExactMatchType:
+			parts = append(parts, fmt.Sprintf(`connection.uri_san_peer_certificate == %s`, strconv.Quote(m.SpiffeID.Value)))
+		case common_api.PrefixMatchType:
+			parts = append(parts, fmt.Sprintf(`connection.uri_san_peer_certificate.startsWith(%s)`, strconv.Quote(m.SpiffeID.Value)))
+		}
+	}
+	if m.SNI != nil && m.SNI.Type == common_api.SNIExactMatchType {
+		parts = append(parts, fmt.Sprintf(`connection.requested_server_name == %s`, strconv.Quote(m.SNI.Value)))
+	}
+	return strings.Join(parts, " && ")
+}
 
 // BuildAccessLogBuildersFromRules turns inbound rules into access log builders.
 // Each rule fires independently: every backend on every rule whose Match holds
